% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesClassifier.R
\name{predict.BayesClassifier}
\alias{predict.BayesClassifier}
\title{Predictions from a BayesClassifier model}
\usage{
\method{predict}{BayesClassifier}(object, newdata, type = "probs", ...)
}
\arguments{
\item{object}{a BayesClassifier object}

\item{newdata}{a data.frame containing new features to predict}

\item{type}{type of return value, either "probs", or "class"}

\item{...}{currently unused}
}
\value{
either a matrix containing log-probabilities, or a vector of classes, see ´details´.
}
\description{
Predicts new data using a BayesClassifier object
}
\details{
Given a BayesClassifier object and a data.frame of new data, predictions are computed for each data point.
Two types of predictions are available:
\itemize{
 \item{´probs´: returns the matrix of log-class probabilities for each sample and class}
 \item{´class´: returns the vector of class labels with the highest predictive probabilities for each sample}
}
}
