% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesClassifier-generics.R
\name{print.BayesClassifier}
\alias{print.BayesClassifier}
\alias{summary.BayesClassifier}
\alias{coef.BayesClassifier}
\alias{dim.BayesClassifier}
\alias{formula.BayesClassifier}
\alias{logLik.BayesClassifier}
\title{BayesClassifier generics}
\usage{
\method{print}{BayesClassifier}(x, ...)

\method{summary}{BayesClassifier}(object, ...)

\method{coef}{BayesClassifier}(object, ...)

\method{dim}{BayesClassifier}(x)

\method{formula}{BayesClassifier}(x, ...)

\method{logLik}{BayesClassifier}(object, ...)
}
\arguments{
\item{x}{a BayesClassifier object}

\item{...}{currently unused}

\item{object}{a BayesClassifier object}
}
\description{
Generic functions implemented for the BayesClassifier class
}
\section{Functions}{
\itemize{
\item \code{print(BayesClassifier)}: prints the BayesClassifier

\item \code{summary(BayesClassifier)}: summarizes the BayesClassifier

\item \code{coef(BayesClassifier)}: returns the list of model parameters in the BayesClassifier

\item \code{dim(BayesClassifier)}: returns the number of classes in the BayesClassifier

\item \code{formula(BayesClassifier)}: returns the formula of the BayesClassifier

\item \code{logLik(BayesClassifier)}: returns the log-likelihood of the BayesClassifier

}}
