% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSC.R
\name{SSC}
\alias{SSC}
\title{Train an SSC-UC model}
\usage{
SSC(
  formula,
  data,
  perc_spies = 0.05,
  naive = FALSE,
  prior = "proportional",
  var_eps = 0.01,
  max_unknown = 6,
  fixed_unknown = NULL,
  runs = 10
)
}
\arguments{
\item{formula}{a formula object explaining the functional relation between input variables and target}

\item{data}{a data.frame containing the dataset}

\item{perc_spies}{percentage of unlabeled points sampled as spies}

\item{naive}{whether a Bayes or Naive Bayes classifier should be used}

\item{prior}{type of class prior; either "uniform" (all classes are equally weighted) or "proportional" (all classes are weighted proportionally to their numbers of samples in the training data)}

\item{var_eps}{scalar to add to the main diagonal of the covariance matrix to assure numerical stability. If 0, no scalar will be added.}

\item{max_unknown}{maximum number of unknown classes (when number of unknown classes is determined automatically)}

\item{fixed_unknown}{fixed number of unknown classes (when number of unknown classes is specified manually)}

\item{runs}{number of bootstrap runs for selecting "likely unknowns"}
}
\value{
a ´BayesClassifier´ object, see \link{BayesClassifier}
}
\description{
Trains a semi-supervised classifier with awareness of unknown classes, as described in \insertCite{Schrunner2020}{SSCUC}.
}
\details{
Given a semi-supervised setup with labeled and unlabeled training data, as well as known and unknown classes (classes represented in the labeled and unlabeled training data or in unlabeled training data only, respectively), a Bayes classifier shall be trained.
The algorithm is described in \insertCite{Schrunner2020}{SSCUC}.
}
\references{
\insertAllCited{}
}
\seealso{
\link{EM} for semi-supervised classification with known classes only, \link{BayesClassifier} for supervised classification
}
