% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesClassifier.R
\name{BayesClassifier}
\alias{BayesClassifier}
\alias{predict.BayesClassifier}
\title{Bayes Classifier or Naive Bayes Classifier}
\usage{
BayesClassifier(formula, data, naive = FALSE, var_eps = 0.01)

\method{predict}{BayesClassifier}(object, newdata, type = "probs", ...)
}
\arguments{
\item{formula}{a formula object explaining the functional relation between input variables and target}

\item{data}{a data.frame containing the dataset}

\item{naive}{whether a Bayes or Naive Bayes classifier should be used}

\item{var_eps}{scalar to add to the main diagonal of the covariance matrix to assure numerical stability. If 0, no scalar will be added.}

\item{object}{a BayesClassifier object}

\item{newdata}{a data.frame containing new features to predict}

\item{type}{type of return value, either "probs", or "class"}

\item{...}{currently unused}
}
\value{
either a matrix containing log-probabilities for each class, or a vector of classes, see 'type'.
}
\description{
Bayes Classifier or Naive Bayes Classifier
}
\section{Methods (by generic)}{
\itemize{
\item \code{predict(BayesClassifier)}: predict new data using a BayesClassifier object

}}
\section{Functions}{
\itemize{
\item \code{BayesClassifier()}: train a BayesClassifier object

}}
\examples{
 set.seed(1)
 dat <- data.frame(y = rep(c(1,2,3), each = 5),
    x1 = as.vector(sapply(c(5,0,0), rnorm, n = 5)),
    x2 = as.vector(sapply(c(0,0,5), rnorm, n = 5)))
 mod <- BayesClassifier(y ~ x1 + x2, dat, naive = TRUE)
 summary(mod)
 predict(mod, newdata = expand.grid(x1 = c(0,5), x2 = c(0,5), y = NA), type = "class")
}
