% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesClassifier.R
\name{BayesClassifier}
\alias{BayesClassifier}
\title{Bayes Classifier or Naive Bayes Classifier}
\usage{
BayesClassifier(
  formula,
  data,
  naive = FALSE,
  prior = "proportional",
  var_eps = 0.01
)
}
\arguments{
\item{formula}{a formula object explaining the functional relation between input variables and target}

\item{data}{a data.frame containing the dataset}

\item{naive}{whether a Bayes or Naive Bayes classifier should be used}

\item{prior}{type of class prior; either "uniform" (all classes are equally weighted) or "proportional" (all classes are weighted proportionally to their numbers of samples in the training data)}

\item{var_eps}{scalar to add to the main diagonal of the covariance matrix to assure numerical stability. If 0, no scalar will be added.}
}
\value{
an S3 object of class ´BayesClassifier´, which has the following internal structure:
\itemize{
 \item{´param´: a list of model parameters; each list element represents one class, and contains a vector mu (class mean), a matrix Sigma (class covariance), and a scalar prior (prior probability)}
 \item{´prior´: the type of prior model used to call the BayesClassifier}
 \item{´formula´: the formula used to call the BayesClassifier}
 \item{´naive´: whether the model contains a Naive Bayes or a full Bayes classifier}
 \item{´n´: the number of samples during training}
 \item{´all.features´: the vector of all feature names in the training data}
 \item{´logLik´: the value of the log-likelihood}
}
}
\description{
Train a BayesClassifier object
}
\details{
The function trains a Bayes classifier on a given classification dataset, specified by a data.frame ´data´ and a formula ´formula´. The specified target variable must be a factor.
The function estimates class-wise mean values (´mu´), covariance matrices (´Sigma´) and prior probabilities (´prior´) to represent classes.
If a Naive Bayes classifier is selected, only a diagonal covariance matrix is estimated.
Prior options indicate whether a uniform prior (all classes are equally weighted), or a proportional prior (all classes are weighted by their proportions in the training data) should be used.
}
\examples{
 set.seed(1)
 dat <- data.frame(y = rep(c(1,2,3), each = 5),
    x1 = as.vector(sapply(c(5,0,0), rnorm, n = 5)),
    x2 = as.vector(sapply(c(0,0,5), rnorm, n = 5)))
 mod <- BayesClassifier(y ~ ., dat, naive = TRUE)
 summary(mod)
 predict(mod, newdata = expand.grid(x1 = c(0,5), x2 = c(0,5), y = NA), type = "class")
}
\seealso{
\link{SSC} for semi-supervised classification with known and unknown classes, \link{EM} for semi-supervised classification with known classes only
}
