% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesClassifier.R
\name{BayesClassifier}
\alias{BayesClassifier}
\title{Train a Gaussian Bayes Classifier}
\usage{
BayesClassifier(formula, data)
}
\arguments{
\item{formula}{a formula object explaining the functional relation between input variables and target}

\item{data}{a data.frame containing the dataset}
}
\description{
Train a Gaussian Bayes Classifier
}
