% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSC.R
\name{EM}
\alias{EM}
\title{Train a semi-supervised Bayes classifier using EM}
\usage{
EM(
  formula,
  data,
  naive = FALSE,
  var_eps = 0.01,
  fixed_labels = NULL,
  verbose = TRUE
)
}
\arguments{
\item{formula}{a formula object explaining the functional relation between input variables and target}

\item{data}{a data.frame containing the dataset}

\item{naive}{whether a Bayes or Naive Bayes classifier should be used}

\item{var_eps}{scalar to add to the main diagonal of the covariance matrix to assure numerical stability. If 0, no scalar will be added.}

\item{fixed_labels}{indices of rows in data, which have fixed labels that must not be changed during training}

\item{verbose}{whether outputs should be shown or suppressed}
}
\value{
a list containing a BayesClassifier object and a vector of labels
}
\description{
Builds and trains a semi-supervised classifier using the Expectation-Maximization (EM) algorithm.
}
