% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesClassifier-generics.R
\name{print.BayesClassifier}
\alias{print.BayesClassifier}
\alias{summary.BayesClassifier}
\alias{coef.BayesClassifier}
\alias{dimnames.BayesClassifier}
\alias{dim.BayesClassifier}
\alias{levels.BayesClassifier}
\alias{length.BayesClassifier}
\alias{nobs.BayesClassifier}
\alias{formula.BayesClassifier}
\alias{logLik.BayesClassifier}
\alias{AIC.BayesClassifier}
\alias{BIC.BayesClassifier}
\title{BayesClassifier generics}
\usage{
\method{print}{BayesClassifier}(x, ...)

\method{summary}{BayesClassifier}(object, ...)

\method{coef}{BayesClassifier}(object, ...)

\method{dimnames}{BayesClassifier}(x)

\method{dim}{BayesClassifier}(x)

\method{levels}{BayesClassifier}(x)

\method{length}{BayesClassifier}(x)

\method{nobs}{BayesClassifier}(object, ...)

\method{formula}{BayesClassifier}(x, ...)

\method{logLik}{BayesClassifier}(object, ...)

\method{AIC}{BayesClassifier}(object, ..., k = 2)

\method{BIC}{BayesClassifier}(object, ...)
}
\arguments{
\item{x}{a BayesClassifier object}

\item{...}{currently unused}

\item{object}{a BayesClassifier object}

\item{k}{penalty term; k=2 in classical AIC}
}
\value{
The form of the returned values of the generic functions for class ´BayesClassifier´ are as follows:
\itemize{
 \item{´print´ and ´summary´ return a console output only}
 \item{´coef´ returns the list of parameters, see \link{BayesClassifier}}
 \item{´dimnames´ and ´levels´ return a vector containing the feature names and class names, respectively}
 \item{´dim´, ´length´ and ´nobs´ return a scalar representing the number of features, the number of classes and the number of training observations, respectively}
 \item{´formula´ returns the formula object used to call the BayesClassifier}
 \item{´logLik´, ´AIC´ and ´BIC´ return the log-likelihood, the AIC and the BIC values, respectively}
}
}
\description{
Generic functions implemented for the BayesClassifier class
}
\section{Functions}{
\itemize{
\item \code{print(BayesClassifier)}: prints the BayesClassifier

\item \code{summary(BayesClassifier)}: summarizes the BayesClassifier

\item \code{coef(BayesClassifier)}: returns the list of model parameters in the BayesClassifier

\item \code{dimnames(BayesClassifier)}: returns the names the features used in the BayesClassifier

\item \code{dim(BayesClassifier)}: returns the dimensionality of the feature space used in the BayesClassifier

\item \code{levels(BayesClassifier)}: returns the names of the classes in the BayesClassifier

\item \code{length(BayesClassifier)}: returns the number of classes in the BayesClassifier

\item \code{nobs(BayesClassifier)}: returns the number of training observations used for the BayesClassifier

\item \code{formula(BayesClassifier)}: returns the formula of the BayesClassifier

\item \code{logLik(BayesClassifier)}: returns the log-likelihood of the BayesClassifier

\item \code{AIC(BayesClassifier)}: returns the Akaike Information Criterion for the BayesClassifier

\item \code{BIC(BayesClassifier)}: returns the Bayesian Information Criterion for the BayesClassifier

}}
