% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSC.R
\name{SSC}
\alias{SSC}
\title{Train an SSC-UC model}
\usage{
SSC(formula, data, perc_spies = 0.05)
}
\arguments{
\item{formula}{a formula object explaining the functional relation between input variables and target}

\item{data}{a data.frame containing the dataset}
}
\value{
a BayesClassifier object
}
\description{
Builds and trains a semi-supervised classifier with awareness of unknown classes, as described in Schrunner et al. (2020).
}
