% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSC.R
\name{SSC}
\alias{SSC}
\title{Train an SSC-UC model}
\usage{
SSC(
  formula,
  data,
  perc_spies = 0.05,
  naive = FALSE,
  var_eps = 0.01,
  max_unknown = 6
)
}
\arguments{
\item{formula}{a formula object explaining the functional relation between input variables and target}

\item{data}{a data.frame containing the dataset}

\item{perc_spies}{percentage of unlabeled points sampled as spies}

\item{naive}{whether a Bayes or Naive Bayes classifier should be used}

\item{var_eps}{scalar to add to the main diagonal of the covariance matrix to assure numerical stability. If 0, no scalar will be added.}

\item{max_unknown}{maximum number of unknown classes}
}
\value{
a BayesClassifier object
}
\description{
Builds and trains a semi-supervised classifier with awareness of unknown classes, as described in Schrunner et al. (2020).
}
